--- a/ARMeilleure/Common/AddressTable.cs
+++ b/ARMeilleure/Common/AddressTable.cs
@@ -4,6 +4,10 @@ using System.Runtime.InteropServices;
 
 namespace ARMeilleure.Common
 {
+    /// <summary>
+    /// Represents a table of guest address to a value.
+    /// </summary>
+    /// <typeparam name="TEntry">Type of the value</typeparam>
     unsafe class AddressTable<TEntry> : IDisposable where TEntry : unmanaged
     {
         public const ulong Mask = ((1ul << 47) - 1) << 2;
@@ -13,6 +17,10 @@ namespace ARMeilleure.Common
         private readonly TEntry _fill;
         private readonly List<IntPtr> _pages;
 
+        /// <summary>
+        /// Gets the base address of the <see cref="EntryTable{TEntry}"/>.
+        /// </summary>
+        /// <exception cref="ObjectDisposedException"><see cref="EntryTable{TEntry}"/> instance was disposed</exception>
         public IntPtr Base
         {
             get
@@ -29,17 +37,34 @@ namespace ARMeilleure.Common
             }
         }
 
+        /// <summary>
+        /// Constructs a new instance of the <see cref="AddressTable{TEntry}"/> class with the specified default fill
+        /// value.
+        /// </summary>
+        /// <param name="fill">Default fill value</param>
         public AddressTable(TEntry fill)
         {
             _fill = fill;
             _pages = new List<IntPtr>(capacity: 16);
         }
 
+        /// <summary>
+        /// Determines if the specified <paramref name="address"/> is mapped on to the table.
+        /// </summary>
+        /// <param name="address">Guest address</param>
+        /// <returns><see langword="true"/> if is mapped; <see langword="false"/> otherwise</returns>
         public bool IsMapped(ulong address)
         {
             return (address & ~Mask) == 0;
         }
 
+        /// <summary>
+        /// Gets a reference to the value at the specified guest <paramref name="address"/>.
+        /// </summary>
+        /// <param name="address">Guest address</param>
+        /// <returns>Reference to the value at the specified guest <paramref name="address"/></returns>
+        /// <exception cref="ObjectDisposedException"><see cref="EntryTable{TEntry}"/> instance was disposed</exception>
+        /// <exception cref="ArgumentException"><paramref name="address"/> is not mapped</exception>
         public ref TEntry GetValue(ulong address)
         {
             if (_disposed)
@@ -58,6 +83,11 @@ namespace ARMeilleure.Common
             }
         }
 
+        /// <summary>
+        /// Gets the leaf page for the specified guest <paramref name="address"/>.
+        /// </summary>
+        /// <param name="address">Guest address</param>
+        /// <returns>Leaf page for the specified guest <paramref name="address"/></returns>
         private TEntry* GetPage(ulong address)
         {
             var level3 = GetRootPage();
@@ -68,6 +98,10 @@ namespace ARMeilleure.Common
             return level0;
         }
 
+        /// <summary>
+        /// Lazily initialize and get the root page of the <see cref="AddressTable{TEntry}"/>.
+        /// </summary>
+        /// <returns>Root page of the <see cref="AddressTable{TEntry}"/></returns>
         private TEntry**** GetRootPage()
         {
             if (_table == null)
@@ -78,6 +112,17 @@ namespace ARMeilleure.Common
             return _table;
         }
 
+        /// <summary>
+        /// Gets the next page at the specified index in the specified table. If the next page is
+        /// <see langword="null"/>, it is initialized to a page of type <typeparamref name="T"/> of the specified
+        /// length.
+        /// </summary>
+        /// <typeparam name="T">Type of the next page</typeparam>
+        /// <param name="level">Current page</param>
+        /// <param name="index">Index in the current page</param>
+        /// <param name="length">Length of the next page</param>
+        /// <param name="fill">Value with which to fill the page if it is initialized</param>
+        /// <returns>Next page</returns>
         private void* GetNextPage<T>(void** level, int index, int length, T fill = default) where T : unmanaged
         {
             ref var result = ref level[index];
@@ -90,6 +135,13 @@ namespace ARMeilleure.Common
             return result;
         }
 
+        /// <summary>
+        /// Allocates a block of memory of the specified type and length.
+        /// </summary>
+        /// <typeparam name="T">Type of elements</typeparam>
+        /// <param name="length">Number of elements</param>
+        /// <param name="fill">Fill value</param>
+        /// <returns>Allocated block</returns>
         private IntPtr Allocate<T>(int length, T fill) where T : unmanaged
         {
             var page = Marshal.AllocHGlobal(sizeof(T) * length);
@@ -102,12 +154,20 @@ namespace ARMeilleure.Common
             return page;
         }
 
+        /// <summary>
+        /// Releases all resources used by the <see cref="AddressTable{TEntry}{TEntry}"/> instance.
+        /// </summary>
         public void Dispose()
         {
             Dispose(true);
             GC.SuppressFinalize(this);
         }
 
+        /// <summary>
+        /// Releases all unmanaged and optionally managed resources used by the <see cref="AddressTable{TEntry}"/>
+        /// instance.
+        /// </summary>
+        /// <param name="disposing"><see langword="true"/> to dispose managed resources also; otherwise just unmanaged resouces</param>
         protected virtual void Dispose(bool disposing)
         {
             if (!_disposed)
@@ -121,6 +181,9 @@ namespace ARMeilleure.Common
             }
         }
 
+        /// <summary>
+        /// Frees resources used by the <see cref="AddressTable{TEntry}"/> instance.
+        /// </summary>
         ~AddressTable()
         {
             Dispose(false);