--- a/ARMeilleure/Translation/DispatcherFunction.cs
+++ b/ARMeilleure/Translation/DispatcherFunction.cs
@@ -1,0 +1,6 @@
+﻿using System;
+
+namespace ARMeilleure.Translation
+{
+    delegate void DispatcherFunction(IntPtr nativeContext, ulong startAddress);
+}