Adds a multi-level function table as described in #2158. ~~The levels are hardcoded into the algorithms, I can make it configurable if this is undesirable.~~

**TODO**
- [x] Add PPTC support.
- [x] Add documentation to `AddressTable<T>`.
- [x] Add fallback for out of range addresses.
- [x] Improve codegen for out of translation branches.

~~Feel free to review/comment even though this is in draft.~~

Closes #2158.