@gdkchan 2021-04-19
> The levels are hardcoded into the algorithms, I can make it configurable if this is undesirable.
Yea, I would prefer if it had some constants for that (it doesn't need to be something that is configurable at runtime, just some compile time constants that we can easily change to control for example, the amount of bits for each level etc).