@FICTURE7 2021-04-19
PPTC could support the shorted table walk path if we stored the guest address in the `PtcIndex` or something, but the change seemed a little intrusive over the PPTC machinery. If its desired I can implement.