@gdkchan 2021-04-19
Btw I think with this change we can enable single block  translation in LowCq (not suggesting doing it as part of this PR, it should probably be a separate change), but it would be nice to know if doing so saves us anything in compilation time and/or code size.