@FICTURE7 2021-04-19
Right with this change LCQ should perform much better and have lower latency, plus we can change the current return behavior to be a virtual jump instead:
https://github.com/Ryujinx/Ryujinx/blob/89791ba68dba70999643c5d876e9329b385c6e8a/ARMeilleure/Instructions/InstEmitFlow.cs#L65-L70
This should allow us to remain in unmanaged for far longer and possibly get rid of this loop:
https://github.com/Ryujinx/Ryujinx/blob/89791ba68dba70999643c5d876e9329b385c6e8a/ARMeilleure/Translation/Translator.cs#L158-L162
Care should be taken around calls though because it will probably stack dive. I guess we can do the virtual jump return only if the call depth is 1 or avoid trying to mimic the guest calls stack by always doing jumps.

However I suspect that if we enable single block decode in LCQ, the JIT code will overlap more, at least with the current rejit policy. Take the following flow for example: `block A -> block B -> block C`. If block A becomes hot, block B and C too will probably become hot, and since presumably HCQ will not be single block decode, an HCQ compile at block A will overlap [A, B, C] and a compile B will overlap [B, C]. I guess we have to figure a new rejit policy or do fancy tracing?