@gdkchan 2021-04-20
> I guess we have to figure a new rejit policy or do fancy tracing?

I guess in a ideal world, the decoder would be able to identify and decode a function perfectly, and then we'd just need to rejit call target addresses. But there are several complications, not only because the decoder might hit the instruction limit or mess up in some other way and not decode the entire function, but also due to things like tail calls, where the call is just a B or BR instruction jump.

I think it will need some way to identify all the blocks of a function, and then only rejit from the functions entry point. Perhaps after changing LCQ to use single blocks, we can find a way to represent the "connection" of the functions on the cache, and build some sort of control flow graph with them. Block A functoin would be connected to B, which would be connected to C etc. A rejit request of C could just walk the graph backwards all the way up to the entry point (block A function) and request a rejit of that instead. I guess I should think a bit more about that, but it does sound like something that could work.

Another thing that we will need to eventually take into account is the removal of LCQ functions (and also of HCQ at some point, for cache invalidation). this change should hopefully make the task a little bit easier, since one no longer needs to remove entries from the jump tables.