@FICTURE7 2021-04-20
Here's another idea which should be closer to what we currently do. If an edge single block becomes hot, do an HCQ decode and record that interval in some data structure. If the single block was already overlapping in 
Although there could be other gotchas to this.