@LDj3SNuD 2021-04-20
About PPTC there is a bug in the code in that the `PtcProfiler.UpdateEntry()` method is never called, so rejits (even across multiple runs) are not profiled.
Also, to avoid repeating code, it might be useful to make the `Translator.RegisterFunction()` method static so that it can also be called from `TranslateFuncs()` (@ Ptc.MakeAndSaveTranslations()). It's also understood that the RegisterFunction() method should remove calls to PtcProfiler methods, so that they can be called from the appropriate locations (always post-translate, because the profiled info must be translate-proof).