@FICTURE7 2021-04-20
Yep, there are other issues with PPTC the integration atm. Will push the changes later today or tomorrow.