@EmulationFanatic 2021-04-21
I can't launch any games with this PR. On first boot they just hang after a moment, and on 2nd boot they crash without any error in the console/log.