@FICTURE7 2021-04-21
> I can't launch any games with this PR. On first boot they just hang after a moment, and on 2nd boot they crash without any error in the console/log.

Looks like I broke the table walk during d310ea666d9b443cff322514a4a2695aa3944b0e and hopefully this is the only thing I broke. :eyes: Thanks a lot for testing this. 👍

Also if the extra complexity of d310ea666d9b443cff322514a4a2695aa3944b0e seems unwarranted or concerning, I can add some extra constraints and simplify it to almost what is was.