@FICTURE7 2021-04-21
> Goffy was testing the branch and noticed that it was a bit slower, which hopefully can be improved with some small changes.

We might be able to do that recoup the performance by switching guest returns to virtual jumps instead.

> Goffy also noted that this branch seems to have random 300ms pauses in smash where everything stops - including the music. At a glance, I'm not sure what that is, but it's something to look into.

Quite odd, it could be because of its emitting more IR putting pressure on the pools and the GC.