@FICTURE7 2021-04-21
I'll break this up and add PPTC support for the shorted table walk later today or tomorrow to this PR, then follow up with other optimization PR(s). Just so this remains manageable and regression are caught early.