@EmulationFanatic 2021-04-21
@FICTURE7 Thanks for fixing the game booting issue. I wanted to share with you the results of some performance testing I did of `94a9e17` vs. master, as I was curious about the potential benefits. Note that all shaders were pre-built for this test as I wanted to rule out any associated FPS drops.

**SSBU 10.0.0 torture test (8 players on Fountain of Dreams)**
PPTC On (full purging of PPTC including .info file before building new PPTC with this commit):
![image](https://user-images.githubusercontent.com/62343878/115614010-5b9ba880-a2aa-11eb-818d-b0ca84db543a.png)

PPTC Off:
![image](https://user-images.githubusercontent.com/62343878/115614220-a9181580-a2aa-11eb-996a-002ac5a7b7f7.png)

**Animal Crossing: New Horizons - 10 minutes idling in-game at the starting point (Leischii's save with a full island)**:
PPTC Off:
![image](https://user-images.githubusercontent.com/62343878/115614567-1c218c00-a2ab-11eb-85d4-22c5ee76d209.png)

At this point I stopped testing as it seems there is a common theme and I felt you should know sooner rather than later.