@gdkchan 2021-04-30
Been testing some games on this branch, and so far I only noticed a significant improvement with ACA NEO GEO.
master:
![image](https://user-images.githubusercontent.com/5624669/116639359-c1d59a80-a93e-11eb-9e3b-efa3cc23a564.png)
PR:
![image](https://user-images.githubusercontent.com/5624669/116639367-c7cb7b80-a93e-11eb-978b-1ed84e23f843.png)
Around +6 FPS ingame on this title.
No regressions observed so far (in performance or games breaking). Good job.