@Leischii 2021-04-30
Regression also seems to be fixed for AC, maybe even slightly better fps now:

PR:
https://i.gyazo.com/thumb/1000/2d63752f04f4225b77e806b76eeda54e-jpg