@EmulationFanatic 2021-05-26
Since this seems to be the final state of the PR, I figured I'd update my numbers.

**SSBU Stats**
Boot time
Master: 48.09s
PR: 47.1s

Avg FPS during torture test:
Master: 52.3
PR: 54.6

**ACNH Stats**
Boot time
Master: 68.58s
PR: 64.91s

Avg FPS during afternoon on Leischii's island:
Master: 59.3
PR: 59.1

These certainly paint the PR's performance in a much better light than last month when I first had compared. Great work!